data_ingestion/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
tqdm==4.66.1
google-cloud-bigquery-storage==2.30.0 